requirements.txt
pandas~=2.1.3
transformers~=4.35.2
torch~=2.1.1
tqdm~=4.66.1
pyspark
pyarrow
grpcio
google-api-python-client
psycopg2-binary
numpy~=1.26.2
nltk~=3.8.1